@@ -26,23 +26,23 @@ version = "2.0.0"
 packages = [{ include = "coral" }]
 
 [tool.poetry.dependencies]
-cvxopt = "*"
-cnvkit = "*"
+cvxopt = ">=1.3.2"
+cnvkit = { git = "https://github.com/suhas-r/cnvkit.git", branch = "master" }
 intervaltree = "*"
 matplotlib = "*"
-pandas = { version = "*", python = ">=3.7,<4.0" }
+pandas = { version = ">=2.2.3", python = "=3.12.*" }
 pysam = { version = ">0.1.7" }
 pytest = { version = ">=4.4", optional = true }
-python = ">=3.10,<4.0"
+python = "=3.12.*"
 numpy = "==2.0.2"
 numba = "==0.60.0"
 llvmlite = "==0.43.0"
-pomegranate = "==1.1.0"
 scanpydoc = { version = ">=0.5", optional = true }
 typing-extensions = ">=3.7.4"
 typing_extensions = { version = "*", python = "<3.9", optional = true }
 click = "^8.1.7"
 mypy = "^1.11.2"
+pomegranate = "^1.1.0" # Py3.12 support
 pydantic = "^2.9.2"
 pyomo = "^6.8.0"
 ruff = "^0.6.9"
@@ -50,6 +50,12 @@ highspy = "^1.7.2"
 pyranges = "^0.1.2"
 seaborn = "^0.13.2"
 typer = "^0.12.5"
+torch = { version =">=2.5.1, =2.*", source="pytorch-cpu"} # Required by pomegranate/cnvkit
+
+[[tool.poetry.source]]
+name = "pytorch-cpu"
+url = "https://download.pytorch.org/whl/cpu"
+priority = "explicit"
 
 [build-system]
 build-backend = "poetry.core.masonry.api"
@@ -83,6 +89,7 @@ lint.select = ["ALL"]
 lint.ignore = [
   "E501",
   "D",       # Docstrings
+  "TD",      # TODO metadata
   "COM812",  # Trailing comma
   "PLR2004", # "magic" value in comparison rather than constant
   "G004",    # f-strings
@@ -92,6 +99,12 @@ lint.ignore = [
   "T201",    # No prints
   "TRY003",  # Long messages outside exception class
   "EM102",   # f-string in exception
+  "ERA001",  # commented-out code, inaccurate
+  "RUF005",  # Prefer iterable unpacking over concatenation
+  "PLR0911", # too-many-return-statements
+  "FIX002",  # Line contains TODO
+  "TRY400",  # logger.error vs. logger.exception (don't want to log traceback)
+  "TCH001",  # typing-only-first-party-import
 ]
 lint.extend-select = ["I"]
 line-length = 80
@@ -100,4 +113,24 @@ lint.pycodestyle.max-doc-length = 80
 [tool.mypy]
 enable_error_code = "possibly-undefined"
 disable_error_code = "import-untyped"
-# strict = true
+exclude = [
+  "hsr.py", 
+  "scripts/report_nanopore_qc.py",
+  "plot_amplicons.py", 
+  "plot_cn.py"
+]
+plugins = "numpy.typing.mypy_plugin"
+strict = false
+
+[tool.pyright]
+exclude = [
+    "**/__pycache__",
+    "**/report_nanopore_qc.py"
+]
+defineConstant = { DEBUG = true }
+
+reportMissingImports = "error"
+reportMissingTypeStubs = false
+
+pythonVersion = "3.12"
+pythonPlatform = "Linux"
