@@ -7,19 +7,26 @@ from __future__ import annotations
 
 import logging
 import warnings
+from collections import defaultdict
 from dataclasses import dataclass, field
-from typing import Any
 
 import cvxopt  # type: ignore[import-untyped]
 import cvxopt.modeling  # type: ignore[import-untyped]
 import numpy as np
 
-from coral import types
+from coral import core_types, datatypes
 from coral.breakpoint.breakpoint_utilities import (
     check_valid_discordant_rc_partition,
     enumerate_partitions,
 )
-from coral.constants import CHR_TAG_TO_IDX
+from coral.datatypes import (
+    AdjacencyMatrix,
+    ConcordantEdge,
+    Node,
+    SequenceEdge,
+    Strand,
+    WalkData,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -28,25 +35,57 @@ logger = logging.getLogger(__name__)
 class BreakpointGraph:
     """A container object for the breakpoint graphs."""
 
-    amplicon_intervals: list[types.AmpliconInterval] = field(
+    amplicon_intervals: list[datatypes.AmpliconInterval] = field(
         default_factory=list
     )
-    sequence_edges: list[list[Any]] = field(default_factory=list)
-    concordant_edges: list[list[Any]] = field(default_factory=list)
-    discordant_edges: list[list[Any]] = field(default_factory=list)
-    source_edges: list[list[Any]] = field(default_factory=list)
+    sequence_edges: list[datatypes.SequenceEdge] = field(default_factory=list)
+    concordant_edges: list[datatypes.ConcordantEdge] = field(
+        default_factory=list
+    )
+    discordant_edges: list[datatypes.DiscordantEdge] = field(
+        default_factory=list
+    )
+    source_edges: list[datatypes.SourceEdge] = field(default_factory=list)
 
     """
-	nodes: adjacent list - keys with format (chr, pos, orientation); 
+	nodes: adjacent list - keys with format Node(chr, pos, orientation);
 	vals = [[sequence edges], [concordant edges], [discordant edges], [source edges]]  
 	"""
-    nodes: dict[tuple, list[list]] = field(default_factory=dict)
-    endnodes: dict[tuple, list[list]] = field(default_factory=dict)
+    node_adjacencies: dict[datatypes.Node, AdjacencyMatrix] = field(
+        default_factory=lambda: defaultdict(AdjacencyMatrix)
+    )
+    endnode_adjacencies: defaultdict[datatypes.Node, list[int]] = field(
+        default_factory=lambda: defaultdict(list)
+    )
+
+    # TODO: add docstring
+    path_constraints: list[datatypes.PathConstraint] = field(
+        default_factory=list
+    )
+    longest_path_constraints: list[datatypes.FinalizedPathConstraint] = field(
+        default_factory=list
+    )
+
+    normal_cov: float = 0.0
     max_cn: float = 0.0
+    amplicon_idx: int = 0
+
+    # Only filled in after valid LP solution
+    # TODO: add docstring
+    walks: WalkData[datatypes.OptimizationWalk] = field(
+        default_factory=lambda: WalkData([], [])
+    )
+    walk_weights: WalkData[float] = field(
+        default_factory=lambda: WalkData([], [])
+    )
+    path_constraints_satisfied: WalkData[list[int]] = field(
+        default_factory=lambda: WalkData([], [])
+    )
 
     @property
     def num_nodes(self) -> int:
-        return len(self.nodes)
+        """Number of nodes in the breakpoint graph (excluding s and t)."""
+        return len(self.node_adjacencies)
 
     @property
     def num_edges(self) -> int:
@@ -55,7 +94,7 @@ class BreakpointGraph:
             + self.num_conc_edges
             + self.num_disc_edges
             + 2 * self.num_src_edges
-            + 2 * len(self.endnodes)
+            + 2 * len(self.endnode_adjacencies)
         )
 
     @property
@@ -78,344 +117,115 @@ class BreakpointGraph:
     def num_nonsrc_edges(self) -> int:
         return self.num_seq_edges + self.num_conc_edges + self.num_disc_edges
 
-    def add_node(self, node_):
-        """Add a new node to the breakpoint graph.
+    @property
+    def total_interval_size(self) -> int:
+        return sum(len(interval) for interval in self.amplicon_intervals)
 
-        Args:
-                node_: A breakpoint node.
+    @property
+    def num_chromosomes(self) -> int:
+        return len({seq.chr for seq in self.sequence_edges})
 
-        """
-        if type(node_) != tuple or len(node_) != 3:
-            raise Exception(
-                "Breakpoint node must be of form (chr, pos, orientation)."
-            )
-        if node_ in self.nodes:
-            pass
-        self.nodes[node_] = [[], [], [], []]
-
-    def add_endnode(self, node_):
+    def add_endnode(self, node_: datatypes.Node) -> None:
         """Add a new node to the list corresponding to interval ends.
 
         Args:
                 node_: A breakpoint node.
 
         """
-        if type(node_) != tuple or len(node_) != 3:
-            raise Exception(
-                "Breakpoint node must be of form (chr, pos, orientation)."
-            )
-        if node_ not in self.endnodes:
-            self.endnodes[node_] = []
+        if node_ not in self.endnode_adjacencies:
+            self.endnode_adjacencies[node_] = []
         else:
             warnings.warn("Node corresponding to interval end already exists.")
 
-    def del_endnode(self, node_):
-        """Delete a node corresponding to interval ends.
-
-        Args:
-                node_: A breakpoint node.
-
-        """
-        if node_ in self.endnodes:
-            del self.endnodes[node_]
-        else:
-            warnings.warn("Node corresponding to interval end not exists.")
-
-    def del_discordant_endnodes(self):
-        """Delete nodes correspond to interval ends and connect to a discordant edges."""
-        del_list = []
-        for node in self.endnodes:
-            if len(self.endnodes[node]) > 0:
-                del_list.append(node)
-        for node in del_list:
-            del self.endnodes[node]
-
     def add_sequence_edge(
-        self, chr, l, r, sr_count=-1, sr_flag="d", lr_count=-1, lr_nc=0, cn=0.0
-    ):
+        self,
+        chr: core_types.ChrTag,
+        l: int,
+        r: int,
+        lr_count: int = -1,
+        lr_nc: float = 0.0,
+        cn: float = 0.0,
+    ) -> None:
         """Add a sequence edge to the graph."""
-        if (chr, l, "-") not in self.nodes or (chr, r, "+") not in self.nodes:
-            raise Exception("Breakpoint node must be added first.")
-        lseq = len(self.sequence_edges)
-        self.nodes[(chr, l, "-")][0].append(lseq)
-        self.nodes[(chr, r, "+")][0].append(lseq)
-        self.sequence_edges.append(
-            [chr, l, r, sr_count, sr_flag, lr_count, lr_nc, r - l + 1, cn]
+        node1, node2 = (
+            Node(chr, l, Strand.REVERSE),
+            Node(chr, r, Strand.FORWARD),
         )
+        lseq = len(self.sequence_edges)
+        self.node_adjacencies[node1].sequence.append(lseq)
+        self.node_adjacencies[node2].sequence.append(lseq)
+        self.sequence_edges.append(SequenceEdge(chr, l, r, lr_nc, lr_count, cn))
 
     def add_concordant_edge(
         self,
-        chr1,
-        pos1,
-        o1,
-        chr2,
-        pos2,
-        o2,
-        sr_count=-1,
-        sr_flag="d",
-        lr_count=-1,
-        reads=set([]),
-        cn=0.0,
-    ):
+        node1: Node,
+        node2: Node,
+        lr_count: int = -1,
+        reads: set[str] | None = None,
+        cn: float = 0.0,
+    ) -> None:
         """Add a concordant edge to the graph."""
-        if chr1 != chr2 or pos2 != pos1 + 1 or o1 != "+" or o2 != "-":
+        if (
+            node1.chr != node2.chr
+            or node2.pos != node1.pos + 1
+            or node1.strand != Strand.FORWARD
+            or node2.strand != Strand.REVERSE
+        ):
             raise Exception("Invalid concordant edge.")
-        if (chr1, pos1, o1) not in self.nodes or (
-            chr2,
-            pos2,
-            o2,
-        ) not in self.nodes:
-            raise Exception("Breakpoint node must be added first.")
+        if reads is None:
+            reads = set()
         lc = len(self.concordant_edges)
-        self.nodes[(chr1, pos1, o1)][1].append(lc)
-        self.nodes[(chr2, pos2, o2)][1].append(lc)
+        self.node_adjacencies[node1].concordant.append(lc)
+        self.node_adjacencies[node2].concordant.append(lc)
         self.concordant_edges.append(
-            [
-                chr1,
-                pos1,
-                o1,
-                chr2,
-                pos2,
-                o2,
-                sr_count,
-                sr_flag,
-                lr_count,
-                reads,
-                cn,
-            ],
+            ConcordantEdge(node1, node2, lr_count, cn, read_names=reads)
         )
 
     def add_discordant_edge(
         self,
-        chr1,
-        pos1,
-        o1,
-        chr2,
-        pos2,
-        o2,
-        sr_count=-1,
-        sr_flag="d",
-        sr_cn=0.0,
-        lr_count=-1,
-        reads=set([]),
-        cn=0.0,
-    ):
+        node1: Node,
+        node2: Node,
+        read_support: int | None = None,
+        cn: float = 0.0,
+        alignments: set[datatypes.BPAlignments] | None = None,
+    ) -> None:
         """Add a discordant edge to the breakpoint graph."""
-        if (chr1, pos1, o1) not in self.nodes or (
-            chr2,
-            pos2,
-            o2,
-        ) not in self.nodes:
-            raise Exception("Breakpoint node must be added first.")
+
+        if not alignments:
+            alignments = set()
+
         ld = len(self.discordant_edges)
-        self.nodes[(chr1, pos1, o1)][2].append(ld)
-        self.nodes[(chr2, pos2, o2)][2].append(ld)
-        if (chr1, pos1, o1) in self.endnodes:
-            self.endnodes[(chr1, pos1, o1)].append(ld)
-        if (chr2, pos2, o2) in self.endnodes:
-            self.endnodes[(chr2, pos2, o2)].append(ld)
+        self.node_adjacencies[node1].discordant.append(ld)
+        self.node_adjacencies[node2].discordant.append(ld)
+        if node1 in self.endnode_adjacencies:
+            self.endnode_adjacencies[node1].append(ld)
+        if node2 in self.endnode_adjacencies:
+            self.endnode_adjacencies[node2].append(ld)
         self.discordant_edges.append(
-            [
-                chr1,
-                pos1,
-                o1,
-                chr2,
-                pos2,
-                o2,
-                sr_count,
-                sr_flag,
-                sr_cn,
-                lr_count,
-                reads,
-                cn,
-            ],
-        )
-
-    def del_discordant_edges(self, del_list, bpi_map):
-        """Delete a list discordant edges from the graph."""
-        sorted_del_list = sorted(del_list, reverse=True)
-        for bpi in sorted_del_list:
-            del self.discordant_edges[bpi]
-        for node in self.endnodes.keys():
-            for i in range(len(self.endnodes[node])):
-                if self.endnodes[node][i] in sorted_del_list:
-                    del self.endnodes[node][i]
-                else:
-                    self.endnodes[node][i] = bpi_map[self.endnodes[node][i]]
-        for node in self.nodes.keys():
-            for i in range(len(self.nodes[node][2])):
-                if self.nodes[node][2][i] in sorted_del_list:
-                    del self.nodes[node][2][i]
-                else:
-                    self.nodes[node][2][i] = bpi_map[self.nodes[node][2][i]]
-
-    def add_source_edge(
-        self,
-        chr1,
-        pos1,
-        o1,
-        sr_count=0,
-        sr_flag="d",
-        sr_cn=0.0,
-        lr_cn=0.0,
-        cn=0.0,
-    ):
-        """Adds a source edge to the graph."""
-        if (chr1, pos1, o1) not in self.nodes:
-            raise Exception("Breakpoint node must be added first.")
-        self.nodes[(chr1, pos1, o1)][3].append(len(self.source_edges))
-        self.source_edges.append(
-            [
-                "source",
-                -1,
-                "-",
-                chr1,
-                pos1,
-                o1,
-                sr_count,
-                sr_flag,
-                sr_cn,
-                lr_cn,
-                cn,
-            ],
-        )
-
-    def del_source_edges(self, del_list, srci_map):
-        """Delete a list source edges from the graph."""
-        sorted_del_list = sorted(del_list, reverse=True)
-        for srci in sorted_del_list:
-            del self.source_edges[srci]
-        for node in self.nodes.keys():
-            for i in range(len(self.nodes[node][3])):
-                if self.nodes[node][3][i] in sorted_del_list:
-                    del self.nodes[node][3][i]
-                else:
-                    self.nodes[node][3][i] = srci_map[self.nodes[node][3][i]]
-
-    def del_redundant_sequence_edges(self):
-        """Delete redundant sequence edges after merging."""
-        if len(self.discordant_edges) == 0:
-            return
-        del_list = []
-        for seqi in range(len(self.sequence_edges)):
-            sseg = self.sequence_edges[seqi]
-            node1 = (sseg[0], sseg[1], "-")
-            node2 = (sseg[0], sseg[2], "+")
-            s1 = (
-                len(self.nodes[node1][1])
-                + len(self.nodes[node1][2])
-                + len(self.nodes[node1][3])
+            datatypes.DiscordantEdge(
+                node1,
+                node2,
+                lr_count=read_support or 0,
+                alignments=alignments,
+                cn=cn,
             )
-            s2 = (
-                len(self.nodes[node2][1])
-                + len(self.nodes[node2][2])
-                + len(self.nodes[node2][3])
-            )
-            if s1 + s2 == 0:
-                del_list.append(seqi)
-        for seqi in del_list[::-1]:
-            ai = self.sequence_edges[seqi][:3]
-            if ai in self.amplicon_intervals:
-                del self.amplicon_intervals[self.amplicon_intervals.index(ai)]
-            node1 = (ai[0], ai[1], "-")
-            node2 = (ai[0], ai[2], "+")
-            del self.sequence_edges[seqi]
-            del self.nodes[node1]
-            del self.nodes[node2]
-            self.del_endnode(node1)
-            self.del_endnode(node2)
-        for seqi in range(len(self.self.sequence_edges)):
-            sseg = self.sequence_edges[seqi]
-            node1 = (sseg[0], sseg[1], "-")
-            node2 = (sseg[0], sseg[2], "+")
-            self.nodes[node1][0][0] = seqi
-            self.nodes[node2][0][0] = seqi
+        )
 
-    def merge_edges(self):
-        """Merge sequence edges connected only by concordant edges;
-        Delete the nodes and concordant edges accordingly.
+    def sort_edges(self) -> None:
+        """Sort sequence and concordant edges according to chromosome and
+        position, then reset adjacency matrix.
         """
-        c_del_list = []  # The list of concordant edges to be deleted
-        seq_del_list = []  # The list of sequence edges to be deleted
-        for ci in range(len(self.concordant_edges)):
-            ce = self.concordant_edges[ci]
-            node1 = (ce[0], ce[1], ce[2])
-            node2 = (ce[3], ce[4], ce[5])
-            if (
-                len(self.nodes[node1][2]) == 0
-                and len(self.nodes[node2][2]) == 0
-                and len(self.nodes[node1][3]) == 0
-                and len(self.nodes[node2][3]) == 0
-            ):
-                seqi1 = self.nodes[node1][0][0]
-                seqi2 = self.nodes[node2][0][0]
-                seq_del_list.append(seqi1)
-                del self.nodes[node1]
-                del self.nodes[node2]
-                c_del_list.append(ci)
-        seq_del_list = sorted(seq_del_list)
-        si = 0
-        li = 0
-        for i in range(1, len(seq_del_list)):
-            if seq_del_list[i] == seq_del_list[li] + 1:
-                li += 1
-            else:
-                seqi1 = seq_del_list[si]
-                seqi2 = seq_del_list[li] + 1
-                self.sequence_edges[seqi2][1] = self.sequence_edges[seqi1][1]
-                self.sequence_edges[seqi2][3] = -1
-                self.sequence_edges[seqi2][4] = "f"
-                self.sequence_edges[seqi2][-2] = (
-                    self.sequence_edges[seqi2][2]
-                    - self.sequence_edges[seqi2][1]
-                    + 1
-                )
-                si = i
-                li = i
-        seqi1 = seq_del_list[si]
-        seqi2 = seq_del_list[li] + 1
-        self.sequence_edges[seqi2][1] = self.sequence_edges[seqi1][1]
-        self.sequence_edges[seqi2][3] = -1
-        self.sequence_edges[seqi2][4] = "f"
-        self.sequence_edges[seqi2][-2] = (
-            self.sequence_edges[seqi2][2] - self.sequence_edges[seqi2][1] + 1
-        )
-        for seqi in seq_del_list[::-1]:
-            del self.sequence_edges[seqi]
-        for ci in sorted(c_del_list, reverse=True):
-            del self.concordant_edges[ci]
-        for seqi in range(len(self.sequence_edges)):
-            sseg = self.sequence_edges[seqi]
-            node1 = (sseg[0], sseg[1], "-")
-            node2 = (sseg[0], sseg[2], "+")
-            self.nodes[node1][0][0] = seqi
-            self.nodes[node2][0][0] = seqi
-        for ci in range(len(self.concordant_edges)):
-            ce = self.concordant_edges[ci]
-            node1 = (ce[0], ce[1], ce[2])
-            node2 = (ce[3], ce[4], ce[5])
-            self.nodes[node1][1][0] = ci
-            self.nodes[node2][1][0] = ci
-
-    def sort_edges(self):
-        """Sort sequence and concordant edges according to chromosome and position
-        Reset adjacent list
-        """
-        self.sequence_edges.sort(
-            key=lambda sseg: (CHR_TAG_TO_IDX[sseg[0]], sseg[1])
-        )
-        self.concordant_edges.sort(
-            key=lambda ce: (CHR_TAG_TO_IDX[ce[0]], ce[1])
-        )
+        self.sequence_edges.sort()
+        self.concordant_edges.sort()
 
         for seqi in range(len(self.sequence_edges)):
             sseg = self.sequence_edges[seqi]
-            self.nodes[(sseg[0], sseg[1], "-")][0] = [seqi]
-            self.nodes[(sseg[0], sseg[2], "+")][0] = [seqi]
+            self.node_adjacencies[sseg.start_node].sequence = [seqi]
+            self.node_adjacencies[sseg.end_node].sequence = [seqi]
         for ci in range(len(self.concordant_edges)):
             ce = self.concordant_edges[ci]
-            self.nodes[(ce[0], ce[1], ce[2])][1] = [ci]
-            self.nodes[(ce[3], ce[4], ce[5])][1] = [ci]
+            self.node_adjacencies[ce.node1].concordant = [ci]
+            self.node_adjacencies[ce.node2].concordant = [ci]
 
     def compute_cn_lr(self, normal_cov_lr: float) -> None:
         """Estimate CN (copy number) for each edge, using only long reads.
@@ -447,34 +257,40 @@ class BreakpointGraph:
         ld = len(self.discordant_edges)
         lsrc = len(self.source_edges)
         logger.debug("Adjacent list for estimating CN:")
-        for node in self.nodes:
-            logger.debug(f"Node {node}; adjacent list = {self.nodes[node]}.")
+        for node in self.node_adjacencies:
+            logger.debug(
+                f"Node {node}; adjacent list = {self.node_adjacencies[node]}."
+            )
         nvariables = lseq + lc + ld + lsrc
         logger.debug(f"There are {nvariables} variables for cvxopt.")
         nconstraints = len(
-            [node for node in self.nodes if node not in self.endnodes]
+            [
+                node
+                for node in self.node_adjacencies
+                if node not in self.endnode_adjacencies
+            ]
         )
         logger.debug(f"There are {nconstraints} constraints for cvxopt.")
 
         wcn = []
         wlncn = []
         wlrseg = []
-        wcn = [0.5 * normal_cov_lr * se[7] for se in self.sequence_edges]
+        wcn = [0.5 * normal_cov_lr * se.gap for se in self.sequence_edges]
         wcn += [normal_cov_lr for eci in range(lc)]
         wcn += [normal_cov_lr for edi in range(ld)]
         wcn += [0.5 * normal_cov_lr for srci in range(lsrc)]
         wlncn = [-0.5 for seg in self.sequence_edges]
-        wlncn += [ce[8] * 1.0 for ce in self.concordant_edges]
-        wlncn += [de[9] * 1.0 for de in self.discordant_edges]
+        wlncn += [ce.lr_count * 1.0 for ce in self.concordant_edges]
+        wlncn += [de.lr_count * 1.0 for de in self.discordant_edges]
         wlncn += [-0.5 for srci in range(lsrc)]
         wlrseg = [
-            (0.5 * se[6] ** 2 / (normal_cov_lr * se[7]))
+            (0.5 * se.lr_nc**2 / (normal_cov_lr * se.gap))
             for se in self.sequence_edges
         ]
         wlrseg += [0.0 for eci in range(lc)]
         wlrseg += [0.0 for edi in range(ld)]
         wlrseg += [
-            (0.5 * self.source_edges[srci][-1] ** 2 / normal_cov_lr)
+            (0.5 * self.source_edges[srci].cn ** 2 / normal_cov_lr)
             for srci in range(lsrc)
         ]
         wcn = cvxopt.matrix(wcn)
@@ -483,15 +299,15 @@ class BreakpointGraph:
 
         cidx = 0
         balance_constraints = np.zeros([nconstraints, nvariables])
-        for node in self.nodes:
-            if node not in self.endnodes:
-                for seqi in self.nodes[node][0]:
+        for node in self.node_adjacencies:
+            if node not in self.endnode_adjacencies:
+                for seqi in self.node_adjacencies[node].sequence:
                     balance_constraints[cidx][seqi] = 1
-                for eci in self.nodes[node][1]:
+                for eci in self.node_adjacencies[node].concordant:
                     balance_constraints[cidx][lseq + eci] = -1
-                for edi in self.nodes[node][2]:
+                for edi in self.node_adjacencies[node].discordant:
                     balance_constraints[cidx][lseq + lc + edi] = -1
-                for srci in self.nodes[node][3]:
+                for srci in self.node_adjacencies[node].source:
                     balance_constraints[cidx][lseq + lc + ld + srci] = -1
                 cidx += 1
         balance_constraints = cvxopt.matrix(balance_constraints)
@@ -535,45 +351,36 @@ class BreakpointGraph:
                     logger.debug(
                         "Reached maximum num iterations.",
                     )
+                logger.debug(f"\tprimal objective = {sol['primal objective']}")
+                logger.debug(f"\tdual objective = {sol['dual objective']}")
+                logger.debug(f"\tgap = {sol['gap']}")
+                logger.debug(f"\trelative gap = {sol['relative gap']}")
                 logger.debug(
-                    "\tprimal objective = %f" % (sol["primal objective"]),
+                    f"\tprimal infeasibility = {sol['primal infeasibility']}"
                 )
                 logger.debug(
-                    "\tdual objective = %f" % (sol["dual objective"]),
-                )
-                logger.debug(
-                    "\tgap = %f" % (sol["gap"]),
-                )
-                logger.debug(
-                    "\trelative gap = %f" % (sol["relative gap"]),
-                )
-                logger.debug(
-                    "\tprimal infeasibility = %f"
-                    % (sol["primal infeasibility"]),
-                )
-                logger.debug(
-                    "dual infeasibility = %f" % (sol["dual infeasibility"]),
+                    f"dual infeasibility = {sol['dual infeasibility']}"
                 )
                 for seqi in range(lseq):
-                    self.sequence_edges[seqi][-1] = sol["x"][seqi] * 2
+                    self.sequence_edges[seqi].cn = sol["x"][seqi] * 2
                     self.max_cn = max(sol["x"][seqi] * 2, self.max_cn)
                 for ci in range(lc):
-                    self.concordant_edges[ci][-1] = sol["x"][lseq + ci] * 2
+                    self.concordant_edges[ci].cn = sol["x"][lseq + ci] * 2
                     self.max_cn = max(sol["x"][lseq + ci] * 2, self.max_cn)
                 for di in range(ld):
                     de = self.discordant_edges[di]
-                    if de[0] == de[3] and de[1] == de[4] and de[2] == de[5]:
-                        self.discordant_edges[di][-1] = sol["x"][lseq + lc + di]
+                    if de.is_self_loop:
+                        self.discordant_edges[di].cn = sol["x"][lseq + lc + di]
                         self.max_cn = max(sol["x"][lseq + lc + di], self.max_cn)
                     else:
-                        self.discordant_edges[di][-1] = (
+                        self.discordant_edges[di].cn = (
                             sol["x"][lseq + lc + di] * 2
                         )
                         self.max_cn = max(
                             sol["x"][lseq + lc + di] * 2, self.max_cn
                         )
                 for srci in range(lsrc):
-                    self.source_edges[srci][-1] = (
+                    self.source_edges[srci].cn = (
                         sol["x"][lseq + lc + ld + srci] * 2
                     )
                     self.max_cn = max(
@@ -584,8 +391,8 @@ class BreakpointGraph:
             logger.debug("Skipped convex optimization.")
             for seqi in range(lseq):
                 se = self.sequence_edges[seqi]
-                cn_seqi = se[6] * 2.0 / (normal_cov_lr * se[7])
-                self.sequence_edges[seqi][-1] = cn_seqi
+                cn_seqi = se.lr_nc * 2.0 / (normal_cov_lr * se.gap)
+                self.sequence_edges[seqi].cn = cn_seqi
                 self.max_cn = max(cn_seqi, self.max_cn)
         self.max_cn += 1.0
 
@@ -606,7 +413,7 @@ class BreakpointGraph:
         Return: a list of integers corresponding to the multiplicities of each
             discordant edge
         """
-        rc_list = [de[9] for de in self.discordant_edges]  # Read counts
+        rc_list = [de.lr_count for de in self.discordant_edges]  # Read counts
         if len(rc_list) == 0:
             return []
         rc_indices = np.argsort(rc_list)
