@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import logging
 from collections import defaultdict
-from typing import Any, Dict, List, Optional
+from typing import Dict, List, Optional
 
 import numpy as np
 import pyomo
@@ -20,7 +20,7 @@ import pyomo.solvers.plugins.solvers.GUROBI
 import pyomo.util.infeasible
 
 from coral import datatypes, models
-from coral.breakpoint import infer_breakpoint_graph
+from coral.breakpoint import breakpoint_graph, infer_breakpoint_graph
 from coral.breakpoint.breakpoint_graph import BreakpointGraph
 from coral.models import cycle_utils, output
 from coral.models.concrete import CycleLPModel
@@ -30,20 +30,18 @@ logger = logging.getLogger(__name__)
 
 
 def minimize_cycles(
-    amplicon_id: int,
     bp_graph: BreakpointGraph,
     k: int,
     total_weights: float,
-    node_order,
-    pc_list,
+    node_order: Dict[datatypes.Node, int],
     p_total_weight=0.9,
     p_bp_cn=0.9,
     solver_options: Optional[datatypes.SolverOptions] = None,
-    solver_to_use: datatypes.Solver = datatypes.Solver.GUROBI,
 ):
     """Cycle decomposition by minimizing the number of cycles/paths.
 
-    Standard model for cycle detection used when generated BreakpointGraph is tractable.
+    Standard model for cycle detection used when generated BreakpointGraph
+    is tractable.
 
     Args:
         amplicon_id: integer, amplicon ID
@@ -51,8 +49,10 @@ def minimize_cycles(
         k: integer, maximum mumber of cycles/paths allowed in cycle decomposition
         total_weights: float, total length-weighted CN in breakpoint graph g
         node_order: dict maps each node in the input breakpoint graph to a distinct integer, indicating a total order of the nodes in g
-        pc_list: list of subpath constraints to be satisfied, each as a dict that maps an edge to its multiplicity
-                *** note that all subpath constraints in this list are required to be satisfied ***
+        pc_list: list of subpath constraints to be satisfied, each as a dict
+            that maps an edge to its multiplicity
+                *** note that all subpath constraints in this list are
+                required to be satisfied ***
                 *** otherwise will return infeasible ***
         p_total_weight: float between (0, 1), minimum proportion of length-weighted CN to be covered by the resulting cycles or paths,
                         default value is 0.9
@@ -64,14 +64,17 @@ def minimize_cycles(
         model_prefix: output prefix for gurobi *.lp model
 
     Returns:
-        A solution container parsed from the MIQCP solver output with the following attributes:
+        A solution container parsed from the MIQCP solver output with the
+        following attributes:
             (1) Pyomo status of optimization model
             (2) Total length weighted CN in resulting cycles/paths
             (3) Total num subpath constraints satisfied by resulting cycles/paths
-            (4) List of cycles, each as a dict which maps an edge to its multiplicity in the cycle
+            (4) List of cycles, each as a dict which maps an edge to its
+                multiplicity in the cycle
             (5) List of the corresponding CN of the above cycles
             (6) Subpath constraints (indices) satisfied by each cycle
-            (7) List of paths, each as a dict which maps an edge to its multiplicity in the path
+            (7) List of paths, each as a dict which maps an edge to its
+                multiplicity in the path
             (8) List of the corresponding CN of the above paths
             (9) Subpath constraints (indices) satisfied by each path
     """
@@ -84,15 +87,15 @@ def minimize_cycles(
         solver_options.time_limit_s, bp_graph.num_disc_edges * 300
     )  # each breakpoint edge is assigned >= 5 minutes)
 
-    model_name = f"amplicon_{amplicon_id}_cycle_decomposition_{k=}"
-    model_filepath = f"{solver_options.output_dir}/{solver_options.model_prefix}_{model_name}"
+    model_name = f"amplicon_{bp_graph.amplicon_idx}_cycle_decomposition_{k=}"
+    prefixed_name = f"{solver_options.model_prefix}_{model_name}"
+    model_filepath = f"{solver_options.output_dir}/{prefixed_name}"
 
     model = models.concrete.get_model(
         bp_graph,
         k,
         total_weights,
         node_order,
-        pc_list,
         model_name=model_name,
     )
 
@@ -115,7 +118,7 @@ def minimize_cycles(
         model,
         bp_graph,
         k,
-        pc_list,
+        bp_graph.longest_path_constraints,
         total_weights,
     )
 
@@ -124,7 +127,7 @@ def minimize_cycles_post(
     amplicon_id: int,
     bp_graph: BreakpointGraph,
     total_weights: float,
-    node_order: Dict[tuple[Any, Any, Any], int],
+    node_order: Dict[datatypes.Node, int],
     pc_list: list,
     init_sol: datatypes.InitialSolution,
     p_total_weight: float = 0.9,
@@ -208,7 +211,6 @@ def minimize_cycles_post(
         k,
         total_weights,
         node_order,
-        pc_list,
         model_name=model_name,
         is_post=True,
         resolution=resolution,
@@ -251,32 +253,34 @@ def initialize_post_processing_solver(
     for i in range(len(init_sol.walks[0])):
         model.z[i] = 1
         model.w[i] = init_sol.walk_weights[0][i]
-        for var_name, var_idx in init_sol.walks[0][i]:
+        for edge_id in init_sol.walks[0][i]:
+            var_name, var_idx = edge_id
             if var_name == "x":
-                model.x[var_idx, i] = init_sol.walks[0][i][(var_name, var_idx)]
+                model.x[var_idx, i] = init_sol.walks[0][i][edge_id]
             elif var_name == "c":
-                model.c[var_idx, i] = init_sol.walks[0][i][(var_name, var_idx)]
+                model.c[var_idx, i] = init_sol.walks[0][i][edge_id]
             elif var_name == "d":
-                model.d[var_idx, i] = init_sol.walks[0][i][(var_name, var_idx)]
+                model.d[var_idx, i] = init_sol.walks[0][i][edge_id]
             elif var_name == "y1":
-                model.y1[var_idx, i] = init_sol.walks[0][i][(var_name, var_idx)]
+                model.y1[var_idx, i] = init_sol.walks[0][i][edge_id]
             elif var_name == "y2":
-                model.y2[var_idx, i] = init_sol.walks[0][i][(var_name, var_idx)]
+                model.y2[var_idx, i] = init_sol.walks[0][i][edge_id]
     for i in range(len(init_sol.walks[1])):
         i_ = i + len(init_sol.walks[0])
         model.z[i_] = 1
         model.w[i_] = init_sol.walk_weights[1][i]
-        for v, vi in init_sol.walks[1][i].keys():
+        for edge_id in init_sol.walks[1][i]:
+            v, vi = edge_id
             if v == "x":
-                model.x[vi, i_] = init_sol.walks[1][i][(v, vi)]
+                model.x[vi, i_] = init_sol.walks[1][i][edge_id]
             elif v == "c":
-                model.c[vi, i_] = init_sol.walks[1][i][(v, vi)]
+                model.c[vi, i_] = init_sol.walks[1][i][edge_id]
             elif v == "d":
-                model.d[vi, i_] = init_sol.walks[1][i][(v, vi)]
+                model.d[vi, i_] = init_sol.walks[1][i][edge_id]
             elif v == "y1":
-                model.y1[vi, i_] = init_sol.walks[1][i][(v, vi)]
+                model.y1[vi, i_] = init_sol.walks[1][i][edge_id]
             elif v == "y2":
-                model.y2[vi, i_] = init_sol.walks[1][i][(v, vi)]
+                model.y2[vi, i_] = init_sol.walks[1][i][edge_id]
     for i in range(len(init_sol.satisfied_pc[0])):
         for pi in init_sol.satisfied_pc[0][i]:
             model.r[pi, i] = 1
@@ -288,20 +292,59 @@ def initialize_post_processing_solver(
             model.R[pi] = 1
 
 
-def maximize_weights_greedy(
-    amplicon_id: int,
+def greedy_solve(
     bp_graph: BreakpointGraph,
     total_weights: float,
-    node_order: Dict[tuple[Any, Any, Any], int],
-    pc_list: List,
+    node_order: Dict[datatypes.Node, int],
+    cycle_id: int,
+    solver_options: datatypes.SolverOptions,
+    alpha: float = 0.01,
+    p_total_weight: float = 0.9,
+    resolution: float = 0.1,
+    cn_tol: float = 0.005,
+    p_subpaths: float = 0.9,
+    *,
+    should_postprocess: bool = False,
+) -> datatypes.CycleSolution:
+    lp_solution = maximize_weights_greedy(
+        bp_graph=bp_graph,
+        total_weights=total_weights,
+        node_order=node_order,
+        cycle_id=cycle_id,
+        alpha=alpha,
+        p_total_weight=p_total_weight,
+        resolution=resolution,
+        cn_tol=cn_tol,
+        p_subpaths=p_subpaths,
+        solver_options=solver_options,
+    )
+    if should_postprocess:
+        lp_solution = postprocess_solution(
+            bp_graph.amplicon_idx,
+            bp_graph,
+            total_weights,
+            node_order,
+            p_total_weight,
+            resolution,
+            lp_solution,
+            solver_options,
+        )
+
+    logger.info("Completed greedy cycle decomposition.")
+    return lp_solution
+
+
+def maximize_weights_greedy(
+    bp_graph: BreakpointGraph,
+    total_weights: float,
+    node_order: Dict[datatypes.Node, int],
     cycle_id: int,
     alpha: float = 0.01,
     p_total_weight: float = 0.9,
     resolution: float = 0.1,
     cn_tol: float = 0.005,
     p_subpaths: float = 0.9,
-    solver_options: Optional[datatypes.SolverOptions] = None,
-    postprocess: int = 0,
+    solver_options: datatypes.SolverOptions | None = None,
 ) -> datatypes.CycleSolution:
     """Greedy cycle decomposition by maximizing the total length-weighted CN.
 
@@ -336,6 +379,8 @@ def maximize_weights_greedy(
     )  # each breakpoint edge is assigned >= 5 minutes)
 
     remaining_weights = total_weights
+
+    pc_list = bp_graph.longest_path_constraints
     is_pc_unsatisfied = [True for i in range(len(pc_list))]
     num_unsatisfied_pc = sum(is_pc_unsatisfied)
     remaining_cn = datatypes.EdgeToCN.from_graph(bp_graph)
@@ -347,7 +392,8 @@ def maximize_weights_greedy(
     )
 
     logger.debug(
-        f"Greedy cycle decomposition with length weighted CN = {remaining_weights} and num subpath constraints = {num_unsatisfied_pc}."
+        f"Greedy cycle decomposition with length weighted CN = "
+        f"{remaining_weights} & num subpath constraints = {num_unsatisfied_pc}."
     )
 
     while next_w >= resolution and (
@@ -356,23 +402,26 @@ def maximize_weights_greedy(
     ):
         pp = 1.0
         if alpha > 0 and num_unsatisfied_pc > 0:
-            pp = (
-                alpha * remaining_weights / num_unsatisfied_pc
-            )  # multi - objective optimization parameter
+            # multi - objective optimization parameter
+            pp = alpha * remaining_weights / num_unsatisfied_pc
         logger.debug(
-            f"Iteration {cycle_id + 1} with remaining CN = {remaining_weights} and unsatisfied constraints = {num_unsatisfied_pc}/{len(pc_list)}."
+            f"Iteration {cycle_id + 1} with remaining CN = {remaining_weights} "
+            f"& unsatisfied constraints = {num_unsatisfied_pc}/{len(pc_list)}."
         )
         logger.debug(f"Multiplication factor for subpath constraints = {pp}.")
 
-        model_name = f"amplicon_{amplicon_id}_cycle_decomposition_greedy_{cycle_id + 1}_{alpha=}"
-        model_filepath = f"{solver_options.output_dir}/{solver_options.model_prefix}_{model_name}"
+        model_name = (
+            f"amplicon_{bp_graph.amplicon_idx}_cycle_decomposition"
+            f"_greedy_{cycle_id + 1}_{alpha=}"
+        )
+        prefixed_name = f"{solver_options.model_prefix}_{model_name}"
+        model_filepath = f"{solver_options.output_dir}/{prefixed_name}"
 
         model = models.concrete.get_model(
             bp_graph,
             k=1,
             total_weights=total_weights,
             node_order=node_order,
-            pc_list=pc_list,
             model_name=model_name,
             is_greedy=True,
             pp=pp,
@@ -429,7 +478,7 @@ def maximize_weights_greedy(
         logger.debug(f"{is_pc_unsatisfied=}")
 
         # Update greedy stop conditions based on latest solution
-        next_w = curr_walk_weight
+        next_w = curr_walk_weight  # type: ignore[possibly-undefined]
         num_unsatisfied_pc = len(pc_list) - len(full_solution.satisfied_pc_set)
         for pi in curr_sol.satisfied_pc_set:
             is_pc_unsatisfied[pi] = False
@@ -443,207 +492,193 @@ def maximize_weights_greedy(
     return full_solution
 
 
-def cycle_decomposition(
-    bb: infer_breakpoint_graph.LongReadBamToBreakpointMetadata,
+def does_graph_require_greedy_solve(bp_graph: BreakpointGraph, k: int) -> bool:
+    return (
+        bp_graph.num_edges > 100
+        or (
+            3 * k
+            + 3 * k * bp_graph.num_edges
+            + 2 * k * bp_graph.num_nodes
+            + k * len(bp_graph.longest_path_constraints)
+        )
+        >= 10000
+    )
+
+
+def solve_single_graph(
+    bp_graph: BreakpointGraph,
+    k: int,
+    total_weights: float,
+    node_order: Dict[datatypes.Node, int],
     solver_options: datatypes.SolverOptions,
     alpha: float = 0.01,
     p_total_weight: float = 0.9,
+    p_bp_cn: float = 0.9,
     resolution: float = 0.1,
-    postprocess: int = 0,
     *,
-    output_all_path_constraints: bool = False,
-) -> None:
-    """Caller for cycle decomposition functions"""
-    was_amplicon_solved: Dict[int, bool] = defaultdict(bool)  # default false
-    for amplicon_idx in range(len(bb.lr_graph)):
-        bp_graph = bb.lr_graph[amplicon_idx]
-        lseg = len(bp_graph.sequence_edges)
-        lc = len(bp_graph.concordant_edges)
-        ld = len(bp_graph.discordant_edges)
-        lsrc = len(bp_graph.source_edges)
-
-        total_weights = 0.0
-        for sseg in bp_graph.sequence_edges:
-            total_weights += sseg[7] * sseg[-1]  # type: ignore[operator]
-        logger.info(f"Begin cycle decomposition for amplicon{amplicon_idx +1}.")
-        logger.info(f"Total CN weights = {total_weights}.")
-
-        bb.longest_path_constraints[amplicon_idx] = longest_path_dict(
-            bb.path_constraints[amplicon_idx],
-        )
-        logger.info(
-            f"Total num maximal subpath constraints = {len(bb.longest_path_constraints[amplicon_idx][0])}."
-        )
-        print(f"Solving {amplicon_idx}")
-        for pathi in bb.longest_path_constraints[amplicon_idx][1]:
-            logger.debug(
-                f"Subpath constraint {pathi} = {bb.path_constraints[amplicon_idx][0][pathi]}"
-            )
-
-        k = max(10, ld // 2)  # Initial num cycles/paths
-        logger.info(f"Initial num cycles/paths = {k}.")
-        nnodes = len(bp_graph.nodes)  # Does not include s and t
-        nedges = lseg + lc + ld + 2 * lsrc + 2 * len(bp_graph.endnodes)
-        node_order = {}
-        ni_ = 0
-        for node in bb.lr_graph[amplicon_idx].nodes:
-            node_order[node] = ni_
-            ni_ += 1
-        if nedges < k:
-            k = nedges
-            logger.info(f"Reset num cycles/paths to {k}.")
-        sol_flag = 0
-        while k <= nedges:
-            if (
-                nedges > 100
-                or (
-                    3 * k
-                    + 3 * k * nedges
-                    + 2 * k * nnodes
-                    + k * len(bb.longest_path_constraints[amplicon_idx][0])
-                )
-                >= 10000
-            ):
-                lp_solution = maximize_weights_greedy(
-                    amplicon_id=amplicon_idx + 1,
-                    bp_graph=bb.lr_graph[amplicon_idx],
-                    total_weights=total_weights,
-                    node_order=node_order,
-                    pc_list=bb.longest_path_constraints[amplicon_idx][0],
-                    cycle_id=0,
-                    alpha=alpha,
-                    p_total_weight=p_total_weight,
-                    resolution=resolution,
-                    cn_tol=0.005,
-                    p_subpaths=0.9,
-                    solver_options=solver_options,
-                    postprocess=postprocess,
-                )
-                if not lp_solution:
-                    logger.info("Greedy cycle decomposition failed.")
-                else:
-                    logger.info("Completed greedy cycle decomposition.")
-                    logger.info(
-                        f"Num cycles = {lp_solution.num_cycles}; num paths = {lp_solution.num_paths}."
-                    )
-                    logger.info(
-                        f"Total length weighted CN = {lp_solution.total_weights_included}/{total_weights}."
-                    )
-                    logger.info(
-                        f"Total num subpath constraints satisfied = {lp_solution.num_pc_satisfied}/{len(bb.longest_path_constraints[amplicon_idx][0])}."
-                    )
-                if postprocess == 1:
-                    lp_solution = postprocess_solution(
-                        amplicon_idx,
-                        bb,
-                        total_weights,
-                        node_order,
-                        p_total_weight,
-                        resolution,
-                        lp_solution,
-                        solver_options,
-                    )
-                bb.walks_by_amplicon[amplicon_idx] = lp_solution.walks
-                bb.walk_weights_by_amplicon[amplicon_idx] = (
-                    lp_solution.walk_weights
-                )
-                bb.path_constraints_satisfied[amplicon_idx] = (
-                    lp_solution.satisfied_pc
-                )
-                sol_flag = 1
-                break
-            lp_solution = minimize_cycles(
-                amplicon_idx + 1,
-                bb.lr_graph[amplicon_idx],
-                k,
-                total_weights,
-                node_order,
-                bb.longest_path_constraints[amplicon_idx][0],
-                p_total_weight,
-                0.9,
-                solver_options,
-            )
-            if (
-                lp_solution.termination_condition
-                == pyo.TerminationCondition.infeasible
-            ):
-                logger.info("Cycle decomposition is infeasible.")
-                logger.info(f"Doubling k from {k} to {k * 2}.")
-                k *= 2
-            else:
-                logger.info(f"Completed cycle decomposition with k = {k}.")
-                logger.info(
-                    f"Num cycles = {len(lp_solution.walks[0])}; num paths = {len(lp_solution.walks[1])}."
-                )
-                logger.info(
-                    f"Total length weighted CN = {lp_solution.total_weights_included}/{total_weights}."
-                )
-                logger.info(
-                    f"Total num subpath constraints satisfied = {len(lp_solution.satisfied_pc_set)}/{len(bb.longest_path_constraints[amplicon_idx][0])}."
-                )
-
-                bb.walks_by_amplicon[amplicon_idx] = lp_solution.walks
-                bb.walk_weights_by_amplicon[amplicon_idx] = (
-                    lp_solution.walk_weights
-                )
-                bb.path_constraints_satisfied[amplicon_idx] = (
-                    lp_solution.satisfied_pc
-                )
-                sol_flag = 1
-                break
-        if sol_flag == 0:
-            logger.info(
-                "Cycle decomposition is infeasible, switch to greedy cycle decomposition."
-            )
-            lp_solution = maximize_weights_greedy(
-                amplicon_id=amplicon_idx + 1,
-                bp_graph=bb.lr_graph[amplicon_idx],
+    should_postprocess: bool = False,
+) -> datatypes.CycleSolution:
+    while k <= bp_graph.num_edges:
+        # When problem is too large, we begin with the greedy optimization
+        if does_graph_require_greedy_solve(bp_graph, k):
+            lp_solution = greedy_solve(
+                bp_graph=bp_graph,
                 total_weights=total_weights,
                 node_order=node_order,
-                pc_list=bb.longest_path_constraints[amplicon_idx][0],
+                solver_options=solver_options,
                 cycle_id=0,
                 alpha=alpha,
                 p_total_weight=p_total_weight,
                 resolution=resolution,
                 cn_tol=0.005,
                 p_subpaths=0.9,
-                solver_options=solver_options,
-                postprocess=postprocess,
+                should_postprocess=should_postprocess,
             )
-            logger.info("Completed greedy cycle decomposition.")
-            logger.info(
-                f"Num cycles = {lp_solution.num_cycles}; num paths = {lp_solution.num_paths}."
-            )
-            logger.info(
-                f"Total length weighted CN = {lp_solution.total_weights_included}/{total_weights}."
-            )
-            logger.info(
-                f"Total num subpath constraints satisfied = {lp_solution.num_pc_satisfied}/{len(bb.longest_path_constraints[amplicon_idx][0])}."
-            )
-            if postprocess == 1:
-                lp_solution = postprocess_solution(
-                    amplicon_idx,
-                    bb,
-                    total_weights,
-                    node_order,
-                    p_total_weight,
-                    resolution,
-                    lp_solution,
-                    solver_options,
-                )
-            bb.walks_by_amplicon[amplicon_idx] = lp_solution.walks
-            bb.walk_weights_by_amplicon[amplicon_idx] = lp_solution.walk_weights
-            bb.path_constraints_satisfied[amplicon_idx] = (
-                lp_solution.satisfied_pc
-            )
-        else:
-            output.output_amplicon_cycles(
-                amplicon_idx,
-                bb,
-                solver_options.output_dir,
-                output_all_path_constraints,
-            )
-            was_amplicon_solved[amplicon_idx] = True
+            break
+
+        lp_solution = minimize_cycles(
+            bp_graph=bp_graph,
+            k=k,
+            total_weights=total_weights,
+            node_order=node_order,
+            p_total_weight=p_total_weight,
+            p_bp_cn=p_bp_cn,
+            solver_options=solver_options,
+        )
+        if (
+            lp_solution.termination_condition
+            != pyo.TerminationCondition.infeasible
+        ):
+            logger.info(f"Completed cycle decomposition with k = {k}.")
+            break
+
+        logger.info("Cycle decomposition is infeasible.")
+        logger.info(f"Doubling k from {k} to {k * 2}.")
+        k *= 2
+        continue
+
+    if lp_solution.termination_condition == pyo.TerminationCondition.infeasible:
+        logger.info(
+            "Cycle decomposition is infeasible, switch to greedy cycle "
+            "decomposition."
+        )
+        lp_solution = greedy_solve(
+            bp_graph=bp_graph,
+            total_weights=total_weights,
+            node_order=node_order,
+            solver_options=solver_options,
+            cycle_id=0,
+            alpha=alpha,
+            p_total_weight=p_total_weight,
+            resolution=resolution,
+        )
+
+    logger.info(
+        f"Num cycles = {lp_solution.num_cycles}; "
+        f"num paths = {lp_solution.num_paths}."
+    )
+    logger.info(
+        f"Total length weighted CN = "
+        f"{lp_solution.total_weights_included}/{total_weights}."
+    )
+    logger.info(
+        f"Total num subpath constraints satisfied = "
+        f"{lp_solution.num_pc_satisfied}/"
+        f"{len(bp_graph.longest_path_constraints)}."
+    )
+    logger.info(f"Completed cycle decomposition with k = {k}.")
+    return lp_solution
+
+
+def cycle_decomposition_single_graph(
+    bp_graph: BreakpointGraph,
+    solver_options: datatypes.SolverOptions,
+    alpha: float = 0.01,
+    p_total_weight: float = 0.9,
+    resolution: float = 0.1,
+    *,
+    should_postprocess: bool = False,
+    output_all_path_constraints: bool = False,
+) -> None:
+    logger.info(
+        f"Begin cycle decomposition for amplicon{bp_graph.amplicon_idx}."
+    )
+
+    total_weights = sum(len(sseg) * sseg.cn for sseg in bp_graph.sequence_edges)
+
+    logger.info(f"Total CN weights = {total_weights}.")
+
+    longest_path_constraints = bp_graph.longest_path_constraints
+
+    logger.info(
+        f"Total num maximal subpath constraints = "
+        f"{len(longest_path_constraints)}."
+    )
+    for pc in longest_path_constraints:
+        logger.debug(
+            f"Subpath constraint {pc.pc_idx} (support={pc.support}) = "
+            f"{bp_graph.path_constraints[pc.pc_idx]}"
+        )
+
+    k = max(10, len(bp_graph.discordant_edges) // 2)
+    logger.info(f"Initial num cycles/paths = {k}.")
+
+    node_order = {
+        node: node_ordered_idx
+        for node_ordered_idx, node in enumerate(bp_graph.node_adjacencies)
+    }
+
+    if bp_graph.num_edges < k:
+        k = bp_graph.num_edges
+        logger.info(f"Reset num cycles/paths to {k}.")
+
+    lp_solution = solve_single_graph(
+        bp_graph=bp_graph,
+        k=k,
+        total_weights=total_weights,
+        node_order=node_order,
+        solver_options=solver_options,
+        alpha=alpha,
+        p_total_weight=p_total_weight,
+        resolution=resolution,
+        should_postprocess=should_postprocess,
+    )
+    bp_graph.walks = lp_solution.walks
+    bp_graph.walk_weights = lp_solution.walk_weights
+    bp_graph.path_constraints_satisfied = lp_solution.satisfied_pc
+
+    output.output_amplicon_walks(
+        bp_graph,
+        solver_options.output_dir,
+        output_all_path_constraints=output_all_path_constraints,
+    )
+
+
+def cycle_decomposition_all_graphs(
+    bb: infer_breakpoint_graph.LongReadBamToBreakpointMetadata,
+    solver_options: datatypes.SolverOptions,
+    alpha: float = 0.01,
+    p_total_weight: float = 0.9,
+    resolution: float = 0.1,
+    *,
+    should_postprocess: bool = False,
+    output_all_path_constraints: bool = False,
+) -> None:
+    """Caller for cycle decomposition functions"""
+    was_amplicon_solved: Dict[int, bool] = defaultdict(bool)  # default false
+    for amplicon_idx in range(len(bb.lr_graph)):
+        bp_graph = bb.lr_graph[amplicon_idx]
+        bp_graph.amplicon_idx = amplicon_idx
+        cycle_decomposition_single_graph(
+            bp_graph,
+            solver_options,
+            alpha,
+            p_total_weight,
+            resolution,
+            should_postprocess=should_postprocess,
+            output_all_path_constraints=output_all_path_constraints,
+        )
+        was_amplicon_solved[amplicon_idx] = True
     output.output_summary_amplicon_stats(
         was_amplicon_solved,
         bb,
@@ -652,56 +687,40 @@ def cycle_decomposition(
 
 
 def reconstruct_cycles(
-    output_dir: str,
-    output_all_path_constraints: bool,
-    cycle_decomp_alpha: float,
-    cycle_decomp_time_limit: int,
-    cycle_decomp_threads: int,
-    solver_to_use: datatypes.Solver,
-    postprocess_greedy_sol: bool,
     bb: infer_breakpoint_graph.LongReadBamToBreakpointMetadata,
-):
+    solver_options: datatypes.SolverOptions,
+    cycle_decomp_alpha: float = 0.01,
+    *,
+    should_postprocess_greedy_sol: bool = False,
+    output_all_path_constraints: bool = False,
+) -> None:
     logging.basicConfig(
-        filename=f"{output_dir}/cycle_decomp.log",
+        filename=f"{solver_options.output_dir}/cycle_decomp.log",
         filemode="w",
-        level=logging.DEBUG,
+        level=logging.INFO,
         format="%(asctime)s:%(levelname)-4s [%(filename)s:%(lineno)d] %(message)s",
     )
-    bb.compute_path_constraints()
     logger.info("Computed all subpath constraints.")
 
-    alpha_ = 0.01
-    postprocess_ = 0
-    nthreads = -1
-    time_limit_ = 7200
-    if cycle_decomp_alpha:
-        alpha_ = cycle_decomp_alpha
-    if postprocess_greedy_sol:
-        postprocess_ = 1
-    if cycle_decomp_threads:
-        nthreads = cycle_decomp_threads
-    if cycle_decomp_time_limit:
-        time_limit_ = cycle_decomp_time_limit
-    cycle_decomposition(
+    cycle_decomposition_all_graphs(
         bb,
-        solver_options=datatypes.SolverOptions(
-            nthreads, time_limit_, output_dir, "pyomo", solver_to_use
-        ),
-        alpha=alpha_,
-        postprocess=postprocess_,
+        solver_options=solver_options,
+        alpha=cycle_decomp_alpha,
+        should_postprocess=should_postprocess_greedy_sol,
         output_all_path_constraints=output_all_path_constraints,
     )
     logger.info("Completed cycle decomposition for all amplicons.")
     logger.info(
-        f"Wrote cycles for all complicons to {output_dir}/amplicon*_cycles.txt."
+        "Wrote cycles for all complicons to "
+        f"{solver_options.output_dir}/amplicon*_cycles.txt."
     )
 
 
 def postprocess_solution(
     amplicon_idx: int,
-    bam_to_bps: infer_breakpoint_graph.LongReadBamToBreakpointMetadata,
+    bp_graph: BreakpointGraph,
     total_weights: float,
-    node_order: Dict[tuple[Any, Any, Any], int],
+    node_order: Dict[datatypes.Node, int],
     p_total_weight: float,
     resolution: float,
     lp_solution: datatypes.CycleSolution,
@@ -717,10 +736,10 @@ def postprocess_solution(
     """
     lp_solution = minimize_cycles_post(
         amplicon_idx + 1,
-        bam_to_bps.lr_graph[amplicon_idx],
+        bp_graph,
         total_weights,
         node_order,
-        bam_to_bps.longest_path_constraints[amplicon_idx][0],
+        bp_graph.longest_path_constraints,
         datatypes.InitialSolution(
             lp_solution.walks,
             lp_solution.walk_weights,
@@ -742,6 +761,8 @@ def postprocess_solution(
         f"Total length weighted CN = {lp_solution.total_weights_included}/{total_weights}."
     )
     logger.info(
-        f"Total num subpath constraints satisfied = {lp_solution.num_pc_satisfied}/{len(bam_to_bps.longest_path_constraints[amplicon_idx][0])}."
+        f"Total num subpath constraints satisfied = "
+        f"{lp_solution.num_pc_satisfied}/"
+        f"{len(bp_graph.longest_path_constraints)}."
     )
     return lp_solution
