@@ -15,15 +15,21 @@ import pyomo.solvers.plugins.solvers
 import pyomo.solvers.plugins.solvers.GUROBI
 import pyomo.util.infeasible
 
-from coral import datatypes, types
+from coral import datatypes
 from coral.breakpoint.breakpoint_graph import BreakpointGraph
-from coral.datatypes import CycleSolution, EdgeToCN
+from coral.datatypes import (
+    CycleSolution,
+    EdgeId,
+    EdgeToCN,
+    EdgeType,
+    OptimizationWalk,
+)
 
 logger = logging.getLogger(__name__)
 
 
 def process_walk_edge(
-    walk: types.AmpliconWalk,
+    walk: OptimizationWalk,
     model: pyo.Model,
     edge_idx: int,
     edge_count: int,
@@ -41,7 +47,7 @@ def process_walk_edge(
 
     # Is sequence edge
     if edge_idx < bp_graph.num_seq_edges:
-        walk[("e", edge_idx)] = edge_count
+        walk[EdgeId(EdgeType.SEQUENCE, edge_idx)] = edge_count
         if remaining_cn:
             remaining_cn.sequence[edge_idx] -= edge_count * model.w[0].value
             if remaining_cn.sequence[edge_idx] < resolution:
@@ -49,7 +55,7 @@ def process_walk_edge(
     # Is concordant edge
     elif edge_idx < bp_graph.num_seq_edges + bp_graph.num_conc_edges:
         conc_edge_idx = edge_idx - bp_graph.num_seq_edges
-        walk[("c", conc_edge_idx)] = edge_count
+        walk[EdgeId(EdgeType.CONCORDANT, conc_edge_idx)] = edge_count
         if remaining_cn:
             remaining_cn.concordant[conc_edge_idx] -= (
                 edge_count * model.w[0].value
@@ -61,7 +67,7 @@ def process_walk_edge(
         disc_edge_idx = (
             edge_idx - bp_graph.num_seq_edges - bp_graph.num_conc_edges
         )
-        walk[("d", disc_edge_idx)] = edge_count
+        walk[EdgeId(EdgeType.DISCORDANT, disc_edge_idx)] = edge_count
         if remaining_cn:
             remaining_cn.discordant[disc_edge_idx] -= (
                 edge_count * model.w[0].value
@@ -74,31 +80,32 @@ def process_walk_edge(
         os.abort()
     # Is source edge
     elif edge_idx < src_node_offset:
-        assert edge_count == 1
         src_edge_idx = edge_idx - bp_graph.num_nonsrc_edges
         if src_edge_idx % 2 == 0:
             s_edge_idx = src_edge_idx // 2
-            walk[("s", s_edge_idx)] = 1  # source edge connected to s
+            # source edge connected to s
+            walk[EdgeId(EdgeType.SOURCE, s_edge_idx)] = 1
             if remaining_cn:
                 remaining_cn.source[s_edge_idx] -= edge_count * model.w[0].value
                 if remaining_cn.source[s_edge_idx] < resolution:
                     remaining_cn.source[s_edge_idx] = 0.0
         else:
             t_edge_idx = (src_edge_idx - 1) // 2
-            walk[("t", t_edge_idx)] = 1  # source edge connected to t
+            # source edge connected to t
+            walk[EdgeId(EdgeType.SINK, t_edge_idx)] = 1
             if remaining_cn:
                 remaining_cn.source[t_edge_idx] -= edge_count * model.w[0].value
                 if remaining_cn.source[t_edge_idx] < resolution:
                     remaining_cn.source[t_edge_idx] = 0.0
+    # Is synthetic end node
+    elif (edge_idx - src_node_offset) % 2 == 0:
+        nsi = (edge_idx - src_node_offset) // 2
+        # source edge connected to s
+        walk[EdgeId(EdgeType.SYNTHETIC_SOURCE, nsi)] = 1
     else:
-        # Is synthetic end node
-        assert edge_count == 1
-        if (edge_idx - src_node_offset) % 2 == 0:
-            nsi = (edge_idx - src_node_offset) // 2
-            walk[("ns", nsi)] = 1  # source edge connected to s
-        else:
-            nti = (edge_idx - src_node_offset - 1) // 2
-            walk[("nt", nti)] = 1  # source edge connected to t
+        nti = (edge_idx - src_node_offset - 1) // 2
+        # source edge connected to t
+        walk[EdgeId(EdgeType.SYNTHETIC_SINK, nti)] = 1
 
 
 def parse_solver_output(
@@ -109,16 +116,16 @@ def parse_solver_output(
     k: int,
     pc_list: List,
     total_weights: float,
-    remaining_cn: Optional[EdgeToCN] = None,
+    remaining_cn: EdgeToCN | None = None,
     resolution: float = 0.0,
     is_pc_unsatisfied: List[bool] | None = None,
 ) -> CycleSolution:
     parsed_sol = CycleSolution(solver_status, solver_termination_condition)
 
     if solver_termination_condition == pyo.TerminationCondition.infeasible:
-        pyomo.util.infeasible.log_infeasible_constraints(
-            model, log_expression=True, log_variables=True
-        )
+        # pyomo.util.infeasible.log_infeasible_constraints(
+        #     model, log_expression=True, log_variables=True
+        # )
         logger.debug("Unable to parse infeasible solution.")
         return parsed_sol
 
@@ -126,63 +133,65 @@ def parse_solver_output(
     lc = len(bp_graph.concordant_edges)
     ld = len(bp_graph.discordant_edges)
     lsrc = len(bp_graph.source_edges)
-    nnodes = len(bp_graph.nodes)  # Does not include s and t
-    nedges = lseg + lc + ld + 2 * lsrc + 2 * len(bp_graph.endnodes)
+    nnodes = len(bp_graph.node_adjacencies)  # Does not include s and t
+    nedges = lseg + lc + ld + 2 * lsrc + 2 * len(bp_graph.endnode_adjacencies)
 
     for i in range(k):
-        logger.debug(f"Walk {i} checking ; CN = {model.w[i].value}.")
-        if model.z[i].value >= 0.9:
-            logger.debug(f"Walk {i} exists; CN = {model.w[i].value}.")
-            if resolution and (walk_weight := model.w[i].value) < resolution:
-                parsed_sol.walk_weights[0].append(walk_weight)
-                logger.debug(
-                    "\tCN less than resolution, iteration terminated successfully."
-                )
+        if model.z[i].value < 0.9:
+            logger.debug(f"Walk {i} does not exist; CN = {model.w[i].value}.")
+            continue
+
+        logger.debug(f"Walk {i} exists; CN = {model.w[i].value}.")
+        if resolution and (walk_weight := model.w[i].value) < resolution:
+            parsed_sol.walk_weights[0].append(walk_weight)
+            logger.debug(
+                "\tCN < resolution, iteration terminated successfully."
+            )
+            break
+        found_cycle = False
+        for node_idx in range(nnodes):
+            if model.c[node_idx, i].value >= 0.9:
+                found_cycle = True
                 break
-            found_cycle = False
-            for node_idx in range(nnodes):
-                if model.c[node_idx, i].value >= 0.9:
-                    found_cycle = True
-                    break
-            cycle: dict = {}
-            path_constraints_s = []
-            for pi in range(len(pc_list)):
-                if model.r[pi, i].value >= 0.9:
-                    path_constraints_s.append(pi)
-                    # Only used for greedy, flip to False when PC satisfied
-                    if is_pc_unsatisfied:
-                        is_pc_unsatisfied[pi] = False
-            for edge_idx in range(nedges):
-                if (edge_count := model.x[edge_idx, i].value) >= 0.9:
-                    edge_count = round(edge_count)
-                    # Update cycle in-place via helper
-                    process_walk_edge(
-                        cycle,
-                        model,
-                        edge_idx,
-                        edge_count,
-                        bp_graph,
-                        is_cycle=found_cycle,
-                        remaining_cn=remaining_cn,
-                        resolution=resolution,
-                    )
-            if (walk_weight := model.w[i].value) > 0.0:
-                if not found_cycle:
-                    parsed_sol.walks[1].append(cycle)
-                    parsed_sol.walk_weights[1].append(walk_weight)
-                    parsed_sol.satisfied_pc[1].append(path_constraints_s)
-                    parsed_sol.satisfied_pc_set |= set(path_constraints_s)
-                else:
-                    parsed_sol.walks[0].append(cycle)
-                    parsed_sol.walk_weights[0].append(walk_weight)
-                    parsed_sol.satisfied_pc[0].append(path_constraints_s)
-                    parsed_sol.satisfied_pc_set |= set(path_constraints_s)
-            for seqi in range(lseg):
-                parsed_sol.total_weights_included += (
-                    model.x[seqi, i].value
-                    * model.w[i].value
-                    * bp_graph.sequence_edges[seqi][-2]
+        cycle: dict = {}
+        path_constraints_s = []
+        for pi in range(len(pc_list)):
+            if model.r[pi, i].value >= 0.9:
+                path_constraints_s.append(pi)
+                # Only used for greedy, flip to False when PC satisfied
+                if is_pc_unsatisfied:
+                    is_pc_unsatisfied[pi] = False
+        for edge_idx in range(nedges):
+            if (edge_count := model.x[edge_idx, i].value) >= 0.9:
+                edge_count = round(edge_count)
+                # Update cycle in-place via helper
+                process_walk_edge(
+                    cycle,
+                    model,
+                    edge_idx,
+                    edge_count,
+                    bp_graph,
+                    is_cycle=found_cycle,
+                    remaining_cn=remaining_cn,
+                    resolution=resolution,
                 )
+        if (walk_weight := model.w[i].value) > 0.0:
+            if not found_cycle:
+                parsed_sol.walks[1].append(cycle)
+                parsed_sol.walk_weights[1].append(walk_weight)
+                parsed_sol.satisfied_pc[1].append(path_constraints_s)
+                parsed_sol.satisfied_pc_set |= set(path_constraints_s)
+            else:
+                parsed_sol.walks[0].append(cycle)
+                parsed_sol.walk_weights[0].append(walk_weight)
+                parsed_sol.satisfied_pc[0].append(path_constraints_s)
+                parsed_sol.satisfied_pc_set |= set(path_constraints_s)
+        for seqi in range(lseg):
+            parsed_sol.total_weights_included += (
+                model.x[seqi, i].value
+                * model.w[i].value
+                * bp_graph.sequence_edges[seqi].gap
+            )
 
     logger.debug(
         f"Total length weighted CN from cycles/paths = {parsed_sol.total_weights_included}/{total_weights}."
